+++ src/geometries/GeometrySet.h
@@ -11,8 +11,8 @@ public:
   GeometrySet(){};
   virtual ~GeometrySet(){};
 
-  virtual bool hitTest(const Ray* ray, float* hitTValue, SurfaceInfo* surfaceInfo) const override;
-  virtual bool hitTest(const Ray* ray, float* hitTValue) const override;
+  virtual bool hitTest(const Ray* ray, double* hitTValue, SurfaceInfo* surfaceInfo) const override;
+  virtual bool hitTest(const Ray* ray, double* hitTValue) const override;
   virtual void constructBoundingBox() override;
 
   virtual bool init() override;
