+++ src/geometries/Mesh.h
@@ -29,8 +29,8 @@ public:
   void setSmoothShading(bool smoothShading);
   void ensureNormalsNormalised();
 
-  virtual bool hitTest(const Ray* ray, float* hitTValue, SurfaceInfo* surfaceInfo) const override;
-  virtual bool hitTest(const Ray* ray, float* hitTValue) const override;
+  virtual bool hitTest(const Ray* ray, double* hitTValue, SurfaceInfo* surfaceInfo) const override;
+  virtual bool hitTest(const Ray* ray, double* hitTValue) const override;
   virtual bool init() override;
 
   virtual const BoundingBoxPtr getBoundingBox() const override;
